library-system/backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
Pillow==11.1.0
django-jazzmin==3.0.1
gunicorn==23.0.0
python-dotenv==1.0.1
whitenoise==6.9.0
